requirements.txt
sentence-transformers
numpy
pandas
apache-airflow